requirements.txt
fastapi
uvicorn
pymongo
pydantic[dotenv]
python-dotenv
transformers
torch
huggingface-hub